requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchaudio
audio-separator
onnxruntime
pydub
soundfile
numpy
pyannote.audio>=3.1
